Fortified reference qualifiers; Hid log stream and log endl classes into log class; Removed non-const iterator support of action bindings.

+++ Engine/Source/System/Window/SystemWindow.cpp
@@ -58,7 +58,10 @@ FSystemWindow::FSystemWindow(std::unique_ptr<ISystemWindowProcedure> WndProc, FS
 
 FSystemWindow::~FSystemWindow() noexcept
 {
-	WndProc->Terminate(*this);
+	if (WndProc != nullptr)
+	{
+		WndProc->Terminate(*this);
+	}
 }
 
 bool FSystemWindow::IsValid() const noexcept
